requirements.txt
git+https://github.com/Pycord-Development/pycord
PyNaCl
pandas
async-timeout
requests
asyncpraw
discordmongopy
motor
dnspython
python-dotenv
discord-components
PycordUtils
